improved EquipmentFactory data structure added FuelTank penalty when digging and moving

+++ GameSession/Actor/Equipment/Modules/DefinesModules.h
@@ -11,7 +11,7 @@ using ModulePrice = hvsdk::CNamedType<unsigned int, struct MODULE_PRICE_TAG>;
 using HullPoints = hvsdk::CNamedType<unsigned int, struct HULL_POINTS_TAG>;
 
 using FuelConsumption = hvsdk::CNamedType<float, struct FUEL_CONSUMPTION_TAG>;
-using FuelAmount = hvsdk::CNamedType<unsigned int, struct FUEL_AMOUNT_TAG>;
+using FuelAmount = hvsdk::CNamedType<float, struct FUEL_AMOUNT_TAG>;
 
 //////////////////////////////////////////////////////////////////////////
